refactor(demo): rename package from sdktest to example
