+++ examples/demo/app/src/huawei/AndroidManifest.xml
@@ -5,7 +5,7 @@
         android:name=".application.MainApplication">
 
         <service
-            android:name="com.onesignal.sdktest.notification.HmsMessageServiceAppLevel"
+            android:name="com.onesignal.example.notification.HmsMessageServiceAppLevel"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.huawei.push.action.MESSAGING_EVENT" />
