+++ examples/demo/app/src/main/AndroidManifest.xml
@@ -10,9 +10,9 @@
     <uses-permission android:name="android.permission.ACCESS_BACKGROUND_LOCATION" />
 
     <permission
-        android:name="com.onesignal.sdktest.permission.RECEIVE_ADM_MESSAGE"
+        android:name="com.onesignal.example.permission.RECEIVE_ADM_MESSAGE"
         android:protectionLevel="signature" />
-    <uses-permission android:name="com.onesignal.sdktest.permission.RECEIVE_ADM_MESSAGE" />
+    <uses-permission android:name="com.onesignal.example.permission.RECEIVE_ADM_MESSAGE" />
     <uses-permission android:name="com.amazon.device.messaging.permission.RECEIVE" />
     <uses-permission android:name="android.permission.WAKE_LOCK" />
 
@@ -48,7 +48,7 @@
             <intent-filter>
                 <action android:name="com.amazon.device.messaging.intent.REGISTRATION" />
                 <action android:name="com.amazon.device.messaging.intent.RECEIVE" />
-                <category android:name="com.onesignal.sdktest" />
+                <category android:name="com.onesignal.example" />
             </intent-filter>
         </receiver>
