+++ examples/demo/app/src/main/java/com/onesignal/example/data/model/NotificationType.kt
@@ -1,0 +1,24 @@
+package com.onesignal.example.data.model
+
+/**
+ * Enum representing different types of push notifications that can be sent.
+ */
+enum class NotificationType(
+    val title: String,
+    val notificationTitle: String,
+    val notificationBody: String,
+    val bigPicture: String? = null,
+    val largeIcon: String? = null
+) {
+    SIMPLE(
+        title = "Simple",
+        notificationTitle = "Simple Notification",
+        notificationBody = "This is a simple push notification"
+    ),
+    WITH_IMAGE(
+        title = "With Image",
+        notificationTitle = "Image Notification",
+        notificationBody = "This notification includes an image",
+        bigPicture = "https://media.onesignal.com/automated_push_templates/ratings_template.png"
+    )
+}
