--- examples/demo/app/src/main/java/com/onesignal/sdktest/ui/components/ToggleRow.kt
@@ -1,79 +1,0 @@
-package com.onesignal.sdktest.ui.components
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.unit.dp
-import com.onesignal.sdktest.ui.theme.OneSignalRed
-
-/**
- * Reusable toggle row with label and optional description.
- * Pass `testTag` and `contentDescription` to match Capacitor's
- * `data-testid` / `aria-label` pattern on `IonToggle`.
- */
-@Composable
-fun ToggleRow(
-    label: String,
-    checked: Boolean,
-    onCheckedChange: (Boolean) -> Unit,
-    modifier: Modifier = Modifier,
-    description: String? = null,
-    enabled: Boolean = true,
-    testTag: String? = null,
-    contentDescription: String? = null,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 14.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = label,
-                style = MaterialTheme.typography.bodyLarge,
-                color = if (enabled) {
-                    MaterialTheme.colorScheme.onSurface
-                } else {
-                    MaterialTheme.colorScheme.onSurface.copy(alpha = 0.4f)
-                }
-            )
-            if (description != null) {
-                Spacer(modifier = Modifier.height(2.dp))
-                Text(
-                    text = description,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(
-                        alpha = if (enabled) 1f else 0.4f
-                    )
-                )
-            }
-        }
-        Spacer(modifier = Modifier.width(12.dp))
-        val switchModifier = Modifier
-            .let { if (testTag != null) it.testTag(testTag) else it }
-            .let { mod ->
-                if (contentDescription != null) {
-                    mod.semantics { this.contentDescription = contentDescription }
-                } else {
-                    mod
-                }
-            }
-        Switch(
-            checked = checked,
-            onCheckedChange = onCheckedChange,
-            enabled = enabled,
-            modifier = switchModifier,
-            colors = SwitchDefaults.colors(
-                checkedThumbColor = MaterialTheme.colorScheme.surface,
-                checkedTrackColor = OneSignalRed,
-                checkedBorderColor = OneSignalRed
-            )
-        )
-    }
-}
